# openapi.yaml
openapi: 3.1.0
...
info:
  title: Oryxa
  version: "1"
  summary: Makes single-user agent frameworks multi-user.
  description: |
    A room holds several people and several agents. Everything that happens is an
    event; sessions and shared context are folds over that log, which is why late
    join, replay, audit and restart recovery are one mechanism rather than four.

    Two things are worth knowing before reading the endpoints.

    **Submitting is not scheduling.** `POST /input` appends one message and
    returns. Who answers, and when, is each agent's own business — the response
    tells you who it woke and why. An agent that stays quiet has not missed it;
    its next turn covers everything it sat through.

    **A connector is data.** `POST /agents` takes the same object you would write
    as YAML, so an agent can be added at runtime with no restart and no
    filesystem access. Registrations are durable and attributed.
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://localhost:8080
    description: A local Oryxa.
security:
  - bearer: []
  - {}   # a server started without -token accepts everything
...
paths:
...
  /v1/agents:
    get:
      tags: [agents]
      summary: Every configured connector.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  agents: { type: array, items: { $ref: "#/components/schemas/Connector" } }
    post:
      tags: [agents]
      summary: Register or replace a connector.
      description: |
        Takes JSON, or YAML with `Content-Type: application/yaml`. Same object
        either way — YAML is how a person writes one on disk, not the interface.

        Durable: the registration is an event, restored on restart and attributed
        to whoever made it. A name already loaded from a file is overridden.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Connector" }
          application/yaml:
            schema: { type: string }
      responses:
        "201":
          description: Registered and live.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Connector" }
        "400": { $ref: "#/components/responses/Invalid" }
  /v1/agents/{name}:
    parameters: [{ $ref: "#/components/parameters/AgentName" }]
    get:
      tags: [agents]
      summary: One connector.
      description: |
        Returns the spec verbatim, headers included. A credential written
        literally into `headers` is therefore readable here — use
        `{{env.NAME}}` so the value stays in the server's environment.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Connector" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [agents]
      summary: Remove a connector.
      description: Sticks across a restart, even for one loaded from a file.
      responses:
        "204": { description: Gone. }
        "404": { $ref: "#/components/responses/NotFound" }
  /v1/agents/{name}/check:
    parameters: [{ $ref: "#/components/parameters/AgentName" }]
    post:
      tags: [agents]
      summary: Probe an agent with a real turn.
      description: |
        Runs the connector against the live agent and reports what came back,
        including the ways a connector can pass while being quietly wrong —
        reasoning leaking into the answer, output arriving twice, selectors that
        matched nothing. This is the Test Connection button.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                probe: { type: string, description: What to say. Defaults to a ping. }
      responses:
        "200":
          description: The agent answered.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CheckResult" }
        "502":
          description: Unreachable, or it answered in a way the connector cannot read.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CheckResult" }
        "404": { $ref: "#/components/responses/NotFound" }
  /v1/sessions:
    get:
      tags: [sessions]
      summary: Every room, newest first.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  sessions: { type: array, items: { $ref: "#/components/schemas/SessionSummary" } }
    post:
      tags: [sessions]
      summary: Open a room.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                agents:
                  type: array
                  items: { type: string }
                  description: Everyone in the room.
                agent:
                  type: string
                  description: A single agent. Kept for callers written before rooms held several.
      responses:
        "201":
          description: Open.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SessionSummary" }
        "400": { $ref: "#/components/responses/Invalid" }
  /v1/sessions/{id}:
    parameters: [{ $ref: "#/components/parameters/SessionID" }]
    get:
      tags: [sessions]
      summary: A room's current state.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SessionView" }
        "404": { $ref: "#/components/responses/NotFound" }
  /v1/sessions/{id}/input:
    parameters: [{ $ref: "#/components/parameters/SessionID" }]
    post:
      tags: [sessions]
      summary: Say something to the room.
      description: |
        Appends one message. It does not create turns — who answers is decided
        here and returned as `wake` and `why`, and the agents woken take their
        turns on their own schedule.

        Leave `to` unset and the room's own rules apply: an agent that was
        @mentioned, named, or has declared an interest in something you said.
        A message naming a person who has spoken here reaches nobody, and neither
        does an acknowledgement like "ok". Anything else reaches everyone.

        `oryxa wake "your message"` answers all of this offline.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [text]
              properties:
                text: { type: string, minLength: 1 }
                author:
                  type: string
                  description: |
                    Who is speaking. Ignored when the server runs with
                    `-trust-header`, which takes it from the proxy instead.
                to:
                  type: array
                  items: { type: string }
                  description: Agents to address explicitly, overriding every rule.
      responses:
        "202":
          description: Accepted.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Input" }
        "400": { $ref: "#/components/responses/Invalid" }
        "401":
          description: The request did not come through the trusted proxy.
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { description: The room is closed. }
  /v1/sessions/{id}/input/{inputId}:
    parameters:
      - { $ref: "#/components/parameters/SessionID" }
      - name: inputId
        in: path
        required: true
        schema: { type: string }
    delete:
      tags: [sessions]
      summary: Withdraw a message nobody has read yet.
      description: |
        Marks rather than removes, so positions never shift. An agent that
        already read it keeps its answer — withdrawing is regret, not time travel.
      responses:
        "204": { description: Withdrawn. }
        "404": { description: No such message, or it was already withdrawn. }
  /v1/sessions/{id}/join:
    parameters: [{ $ref: "#/components/parameters/SessionID" }]
    post:
      tags: [sessions]
      summary: Exchange a room secret for a cookie.
      description: >
        Every other route under /v1/sessions/{id} needs the room secret, sent as
        X-Oryxa-Session. This one takes it in the body and sets a cookie scoped to
        that room instead, because EventSource cannot send a header and the live
        stream has to authenticate too. It is the one route under a session that
        the secret does not already guard — guarding it with the cookie it exists
        to set would lock the door from the inside.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [secret]
              properties:
                secret:
                  type: string
                  description: The secret returned once when the room was created.
      responses:
        "200": { description: Joined; the cookie is set. }
        "404": { $ref: "#/components/responses/NotFound" }
  /v1/sessions/{id}/keys:
    parameters: [{ $ref: "#/components/parameters/SessionID" }]
    post:
      tags: [sessions]
      summary: Issue a key bound to a name.
      description: >
        Returns a room key that carries a name. Presenting it *is* being that
        person: the author on every later request is taken from the key rather
        than from the body, so a key issued for one name cannot be used under
        another. Requires the room secret, which is the right root — whoever
        decides somebody may be in the room decides what to call them. Oryxa
        still does not establish who anyone is; it binds a name to a credential
        and refuses to let the two come apart. Shown once. Reissuing for a name
        replaces the previous key.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [author]
              properties:
                author:
                  type: string
                  description: The name this key speaks as.
      responses:
        "201":
          description: Issued.
          content:
            application/json:
              schema:
                type: object
                properties:
                  session: { type: string }
                  author:  { type: string }
                  key:
                    type: string
                    description: Shown once; not recoverable afterwards.
        "400": { $ref: "#/components/responses/Invalid" }
        "404": { $ref: "#/components/responses/NotFound" }
  /v1/sessions/{id}/cancel:
    parameters: [{ $ref: "#/components/parameters/SessionID" }]
    post:
      tags: [sessions]
      summary: Stop whatever is running.
      description: Abandons the in-flight request. A cancelled turn writes nothing to context.
      responses:
        "202": { description: Cancelling. }
        "404": { $ref: "#/components/responses/NotFound" }
  /v1/sessions/{id}/close:
    parameters: [{ $ref: "#/components/parameters/SessionID" }]
    post:
      tags: [sessions]
      summary: Close the room.
      responses:
        "204": { description: Closed. }
        "404": { $ref: "#/components/responses/NotFound" }
  /v1/sessions/{id}/context:
    parameters: [{ $ref: "#/components/parameters/SessionID" }]
    get:
      tags: [context]
      summary: What the room holds.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  context: { type: array, items: { $ref: "#/components/schemas/ContextEntry" } }
        "404": { $ref: "#/components/responses/NotFound" }
  /v1/sessions/{id}/context/{key}:
    parameters:
      - { $ref: "#/components/parameters/SessionID" }
      - { $ref: "#/components/parameters/ContextKey" }
    post:
      tags: [context]
      summary: Write to shared context.
      description: |
        Send `append` for an add-only list, or `value` for a single mutable
        value. An entry is one or the other for life; the wrong one is a 409.

        A value write may carry `If-Match: <version>`. Without it the write
        overwrites. With it, a stale write is refused **with what is current**,
        so you can merge rather than re-read and guess.
      parameters:
        - name: If-Match
          in: header
          schema: { type: integer, format: int64 }
          description: The version you last saw. Value writes only.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                append: { type: string, description: Add to an add-only list. }
                value: { type: string, description: Replace a single value. }
                author: { type: string }
              oneOf:
                - required: [append]
                - required: [value]
      responses:
        "200":
          description: Written.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ContextEntry" }
        "400": { $ref: "#/components/responses/Invalid" }
        "409":
          description: |
            A stale value write, or the wrong kind for this key. Carries what is
            current so the caller can merge.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Conflict" }
  /v1/sessions/{id}/context/{key}/pin:
    parameters:
      - { $ref: "#/components/parameters/SessionID" }
      - { $ref: "#/components/parameters/ContextKey" }
    post:
      tags: [context]
      summary: Pin or unpin an entry.
      description: |
        Pinned entries are what `{{context.pinned}}` renders — the curated set
        meant to stay small enough to put in front of a model as a room grows.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                pinned: { type: boolean, default: true }
                author: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ContextEntry" }
        "404": { $ref: "#/components/responses/NotFound" }
  /v1/sessions/{id}/events:
    parameters:
      - { $ref: "#/components/parameters/SessionID" }
      - { $ref: "#/components/parameters/Since" }
    get:
      tags: [log]
      summary: The log, from a point.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  events: { type: array, items: { $ref: "#/components/schemas/Event" } }
        "404": { $ref: "#/components/responses/NotFound" }
  /v1/sessions/{id}/stream:
    parameters:
      - { $ref: "#/components/parameters/SessionID" }
      - { $ref: "#/components/parameters/Since" }
    get:
      tags: [log]
      summary: Replay, then follow, over SSE.
      description: |
        `?since=0` replays the whole room and then follows live. Late join,
        reconnect and replay are the same code path, which is the payoff of the
        log being the source of truth rather than a side effect.

        Frames carry no `event:` field, so `EventSource.onmessage` fires for all
        of them. `id:` carries the sequence number, so a reconnect resumes with
        `Last-Event-ID` and loses nothing.
      responses:
        "200":
          description: An event stream.
          content:
            text/event-stream:
              schema: { type: string }
...
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      description: |
        The shared token, if the server was started with one. Off by default,
        and the server says so at startup.
  parameters:
    SessionID:
      name: id
      in: path
      required: true
      schema: { type: string, examples: [s_2f1a9c4e] }
    AgentName:
      name: name
      in: path
      required: true
      schema: { type: string, examples: [langgraph] }
...
  responses:
    NotFound:
      description: No such thing.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Invalid:
      description: The request could not be used.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...